backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "google-genai>=1.70.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
